fix: prevent /P permissions overflow from corrupting decryption key
PdfNumber#intValue() narrowed the underlying double with a plain (int)
cast. Some PDF producers write the /Encrypt dictionary's /P (permissions)
entry using its unsigned 32-bit decimal representation instead of the
equivalent signed value (e.g. 4294965956 instead of -1340) - both encode
the identical 4-byte value, and this is spec-legal.

Java's narrowing double-to-int conversion saturates rather than wraps for
an out-of-range value, so such a /P value was misread as
Integer.MAX_VALUE instead of -1340. That corrupted the encryption key the
standard security handler derives, so its own recomputed /U check failed
for every password - including a correct, empty one - and PdfReader
reported BadPasswordException for a document that was never actually
password-protected.

Route the cast through long first, which reproduces the expected
low-order-32-bit truncation instead of clamping. Applied identically to
both the legacy (com.lowagie) and modern (org.openpdf) copies of
PdfNumber, since this branch ships both.

Add PdfNumberTest covering the truncation directly, and
DecryptUnsignedPermissionsTest, a PdfReader-level regression test using a
crafted PDF whose /P is written this way with an empty user password, to
both modules.

+++ openpdf-core-modern/src/main/java/org/openpdf/text/pdf/PdfNumber.java
@@ -135,11 +135,20 @@ public class PdfNumber extends PdfObject implements Comparable<PdfNumber> {
 
     /**
      * Returns the primitive <CODE>int</CODE> value of this object.
+     * <p>
+     * Some PDF producers write 32-bit values that are meant to be interpreted as a two's-complement
+     * signed integer (for example the {@code /P} entry of an encryption dictionary) using their
+     * equivalent unsigned decimal representation instead of the negative signed value, e.g.
+     * {@code 4294965956} instead of {@code -1340}. A plain {@code (int) value} narrowing cast on the
+     * underlying <CODE>double</CODE> would clamp such an out-of-range value to
+     * {@link Integer#MAX_VALUE} rather than wrapping it, silently corrupting values like that. Going
+     * through <CODE>long</CODE> first reproduces the C-style truncation to the low-order 32 bits that
+     * PDF producers/consumers expect.
      *
      * @return The value as <CODE>int</CODE>
      */
     public int intValue() {
-        return (int) value;
+        return (int) (long) value;
     }
 
     /**
